Move GASP code into separate files
